# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "api",
    "description": "",
    "license": {
      "name": ""
    },
    "version": "0.1.0"
  },
  "paths": {
    "/guilds/{guild_id}": {
      "get": {
        "tags": [
          "crate::routes::guild_config"
        ],
        "operationId": "get_guild_config",
        "parameters": [
          {
            "name": "guild_id",
            "in": "path",
            "description": "",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GuildConfig"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      },
      "patch": {
        "tags": [
          "crate::routes::guild_config"
        ],
        "operationId": "update_guild",
        "parameters": [
          {
            "name": "guild_id",
            "in": "path",
            "description": "",
            "required": true
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GuildUpdates"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": ""
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/oauth/initiate": {
      "get": {
        "tags": [
          "crate::routes::oauth2"
        ],
        "operationId": "initiate",
        "responses": {
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/session/sign-out": {
      "post": {
        "tags": [
          "crate::routes::session"
        ],
        "operationId": "sign_out",
        "responses": {
          "200": {
            "description": ""
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/users/@maybe-me": {
      "get": {
        "tags": [
          "crate::routes::users"
        ],
        "operationId": "get_maybe_user_me",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "type": "null"
                    },
                    {
                      "$ref": "#/components/schemas/FluxerUser"
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/users/@me": {
      "get": {
        "tags": [
          "crate::routes::users"
        ],
        "operationId": "get_user_me",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FluxerUser"
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/users/@me/guilds": {
      "get": {
        "tags": [
          "crate::routes::users"
        ],
        "operationId": "get_user_guilds",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Guild"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "FluxerUser": {
        "type": "object",
        "required": [
          "discriminator",
          "id",
          "is_staff",
          "username"
        ],
        "properties": {
          "avatar": {
            "type": [
              "string",
              "null"
            ]
          },
          "discriminator": {
            "type": "string"
          },
          "global_name": {
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "type": "string"
          },
          "is_staff": {
            "type": "boolean"
          },
          "username": {
            "type": "string"
          }
        }
      },
      "Guild": {
        "type": "object",
        "required": [
          "id",
          "name",
          "owner_id"
        ],
        "properties": {
          "banner": {
            "type": [
              "string",
              "null"
            ]
          },
          "icon": {
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "type": "string"
          },
          "member_count": {
            "type": [
              "integer",
              "null"
            ],
            "minimum": 0
          },
          "name": {
            "type": "string"
          },
          "online_count": {
            "type": [
              "integer",
              "null"
            ],
            "minimum": 0
          },
          "owner_id": {
            "type": "string"
          },
          "splash": {
            "type": [
              "string",
              "null"
            ]
          },
          "vanity_url_code": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
      "GuildConfig": {
        "type": "object",
        "required": [
          "command_prefixes"
        ],
        "properties": {
          "command_prefixes": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
      "GuildUpdates": {
        "type": "object",
        "properties": {
          "command_prefixes": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "string"
            }
          }
        }
      }
    }
  }
}